Support BCP47 language subtags
The `navigator.languages` array and the `navigator.language` property
contain BCP47 language tags that can be parsed more accurately than
using `indexOf`.
Also, the `href` is always built using the same pattern, so that it can
simply be constructed using the known language tags.

Note: This could be written more efficiently using ES6+ syntax, but the
code was kept simple to retain compatibility with older browsers.